+++ tests/unit/ProgressTest.php
@@ -3,7 +3,7 @@
 namespace Test\CpChart;
 
 use Codeception\Test\Unit;
-use CpChart\Chart\Image;
+use CpChart\Image;
 use UnitTester;
 
 class ProgressTest extends Unit