+++ tests/unit/StackedAreaTest.php
@@ -3,8 +3,8 @@
 namespace Test\CpChart;
 
 use Codeception\Test\Unit;
-use CpChart\Chart\Data;
-use CpChart\Chart\Image;
+use CpChart\Data;
+use CpChart\Image;
 use UnitTester;
 
 class StackedAreaTest extends Unit